@@ -1,4 +1,10 @@
-// eslint.config.js
-const antfu = require('@antfu/eslint-config').default
+import antfu from '@antfu/eslint-config'
 
-module.exports = antfu()
+export default antfu({
+  ignores: [
+    'dist',
+    'public',
+    '.github',
+    '.vscode',
+  ],
+})
